moved debts to app router
