+++ src/components/ui/BlurImage.tsx
@@ -1,3 +1,4 @@
+"use client"
 // import Image from "next/image";
 import Image from "next/legacy/image";
 import { useState } from "react";
